Alpha 1 - A few bug fixes, minor changes.
I have decided this commit is going to be called Alpha 1. Currently the
game takes about an hour of continuous play to reach level 25 and obtain
maximum stat equipment. The Admin section of the game is under
construction. I think that's what I'm working on next. I still haven't
implemented Special Encounters properly, so that's on the list, too.
Data binding for the Inventory, Blackjack, and Shop Windows are still on
the list.