Fixed a bug allowing users to purchase items without having the gold
If a user selected an item in The Armoury or The Tavern Bar in a category other than the one they originally purchased an item with, they could still purchase that item in the other category because all purchase buttons weren't checked on purchase. Now all Purchase buttons on all market pages are checked on all purchase/sell.